foundry/foundry_agents.py
"""
AlertEA — Microsoft Foundry Agent Definitions
================================================
Five persistent agents using the agent_framework SDK,
scoped to the 5 divisions of Kampala, Uganda.
"""

import os
from dotenv import load_dotenv
from azure.identity.aio import AzureCliCredential
from agent_framework.azure import AzureAIAgentClient

load_dotenv()

ENDPOINT = os.environ["AZURE_AI_PROJECT_ENDPOINT"]
MODEL    = os.environ["AZURE_AI_MODEL_DEPLOYMENT_NAME"]


def _make_client(credential) -> AzureAIAgentClient:
    return AzureAIAgentClient(
        project_endpoint=ENDPOINT,
        model_deployment_name=MODEL,
        credential=credential,
    )


def weather_agent(credential):
    return _make_client(credential).as_agent(
        name="WeatherAgent",
        instructions="""
        You are WeatherAgent for AlertEA, a disaster early warning
        system for the 5 divisions of Kampala, Uganda
        (Central, Kawempe, Makindye, Nakawa, Rubaga).

        When given weather readings (rainfall mm/hr, wind speed km/h,
        humidity %), you must:
        1. Assess severity of rainfall (low/moderate/high/extreme)
        2. Flag if wind speed exceeds 60 km/h as a secondary hazard
        3. Return ONLY a JSON object with keys:
           rainfall_risk (0-10), wind_risk (0-10), summary (string)

        Be concise. Lives depend on speed.
        """,
    )


def seismic_agent(credential):
    return _make_client(credential).as_agent(
        name="SeismicAgent",
        instructions="""
        You are SeismicAgent for AlertEA. You analyze earthquake data
        from the USGS feed for the 5 divisions of Kampala, Uganda.

        When given seismic readings (magnitude, depth_km, distance_km):
        1. Score seismic risk 0-10: risk = min(10, max(0, (magnitude - 3.0) * 2.5))
        2. Increase score by 20% if depth < 20km (shallow quake)
        3. Return ONLY JSON: {seismic_risk, magnitude, depth_km, summary}

        If no earthquake data, return seismic_risk: 0.
        """,
    )


def flood_agent(credential):
    return _make_client(credential).as_agent(
        name="FloodRiskAgent",
        instructions="""
        You are FloodRiskAgent for AlertEA. Compute flood risk per Kampala division.

        Terrain vulnerability (built-in):
        - Central  (KLA-CEN): 0.80 — Nakivubo wetlands, Kinawataka channel
        - Kawempe  (KLA-KAW): 0.92 — Lubigi wetlands, very low-lying
        - Makindye (KLA-MAK): 0.75 — Ggaba road valleys, Nakisunga channel
        - Nakawa   (KLA-NAK): 0.65 — Murchison Bay shores, Kyambogo slopes
        - Rubaga   (KLA-RUB): 0.82 — Nalukolongo channel, Kasubi slopes

        Formula: flood_score = base * (0.5 + terrain); cap at 10.0
        Base: <10mm=1.0, <25mm=3.5, <40mm=6.0, <60mm=8.0, 60mm+=9.5

        Return ONLY JSON: {flood_risk, terrain_vulnerability, summary}
        """,
    )


def orchestrator_agent(credential):
    return _make_client(credential).as_agent(
        name="OrchestratorAgent",
        instructions="""
        You are OrchestratorAgent for AlertEA — the central intelligence
        of a disaster early warning system for Kampala, Uganda.

        Weights: flood_risk=55%, seismic_risk=30%, wind_risk=15%
        composite = (flood*0.55) + (seismic*0.30) + (wind*0.15)
        Levels: 0-2.5=LOW, 2.5-5=MEDIUM, 5-7.5=HIGH, 7.5-10=CRITICAL
        Primary threat = whichever input score is highest.

        Return ONLY JSON:
        {composite_score, risk_level, primary_threat, recommended_action, confidence_pct}
        """,
    )


def dispatch_agent(credential):
    return _make_client(credential).as_agent(
        name="AlertDispatchAgent",
        instructions="""
        You are AlertDispatchAgent for AlertEA. Write SMS alerts for
        residents in at-risk Kampala divisions.

        Rules:
        1. Maximum 160 characters
        2. Start: 🔴 CRITICAL | 🟠 HIGH | 🟡 MEDIUM | 🟢 LOW
        3. Include: division name, threat, score, action
        4. End with: alertea.africa
        5. Be direct. Lives are at stake.

        Return ONLY the SMS text string.
        """,
    )
