requirements.txt
# AlertEA — Python Dependencies
# Kampala Divisions Disaster Early Warning System

# Web framework
fastapi==0.135.1
uvicorn==0.41.0
starlette==1.0.0rc1

# HTTP client (async)
httpx==0.28.1
httpx-sse==0.4.3
anyio==4.9.0
h11==0.16.0

# Microsoft Agent Framework
agent-framework==1.0.0rc4
agent-framework-core==1.0.0rc4
agent-framework-azure-ai==1.0.0rc4

# Azure AI Foundry
azure-ai-projects==2.0.1
openai==2.26.0

# Azure services
azure-core==1.38.2
azure-identity==1.26.0b2
azure-communication-sms==1.0.1
azure-storage-blob==12.28.0
azure-common==1.1.28

# Data validation
pydantic==2.13.0b2
pydantic_core==2.42.0
pydantic-settings==2.13.1

# Environment
python-dotenv==1.2.2
